Move internal/api/* to api/*
