+++ pkg/csp/chan.go
@@ -6,7 +6,7 @@ import (
 
 	"capnproto.org/go/capnp/v3"
 	casm "github.com/wetware/casm/pkg"
-	api "github.com/wetware/ww/internal/api/channel"
+	api "github.com/wetware/ww/api/channel"
 )
 
 var (
