+++ pkg/host/host.go
@@ -9,11 +9,11 @@ import (
 
 	casm "github.com/wetware/casm/pkg"
 	"github.com/wetware/casm/pkg/cluster"
-	anchor_api "github.com/wetware/ww/internal/api/anchor"
-	api "github.com/wetware/ww/internal/api/cluster"
-	process_api "github.com/wetware/ww/internal/api/process"
-	pubsub_api "github.com/wetware/ww/internal/api/pubsub"
-	reg_api "github.com/wetware/ww/internal/api/registry"
+	anchor_api "github.com/wetware/ww/api/anchor"
+	api "github.com/wetware/ww/api/cluster"
+	process_api "github.com/wetware/ww/api/process"
+	pubsub_api "github.com/wetware/ww/api/pubsub"
+	reg_api "github.com/wetware/ww/api/registry"
 	"github.com/wetware/ww/pkg/anchor"
 	"github.com/wetware/ww/pkg/csp"
 	"github.com/wetware/ww/pkg/pubsub"
