+++ pkg/pubsub/subscription.go
@@ -5,7 +5,7 @@ import (
 
 	"capnproto.org/go/capnp/v3/exp/bufferpool"
 	casm "github.com/wetware/casm/pkg"
-	api "github.com/wetware/ww/internal/api/pubsub"
+	api "github.com/wetware/ww/api/pubsub"
 )
 
 // Subscription is a stateful iterator over a stream of topic messages.
